Upgrade userland to use Debian Buster and updated software
Summary:
This commit allow the PHP nodes to work on
a refreshed environment based on Debian Buster,
OpenJDK 11, PHP 7.4, PHPUnit 9 and phan 3.

It prunes phantomjs, now discontinued in favour
of headless browsers.

Backports aren't enabled, as OpenJDK is directly
available in this Debian version.

Update misc PHP tools versions.

This is a follow-up for 8b10c0ea630b.

Test Plan: Build image, run it on Dwellers.

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2350